small refactor and compile
