+++ src/model/ExtendedGraphQLObjectType.ts
